+++ workflux/users/manage.py
@@ -2,7 +2,7 @@ import requests
 import pprint
 import json
 
-from workfluxlux import db_connector
+from workflux import db_connector
 from flask import current_app as app
 from getpass import getpass
 from time import sleep
