+++ workflux/web_app/static/create_job.js
