+++ workflux/web_app/static/utils.js
